migrate app to new version
