--- app/src/main/java/com/dreamso/radio/SplashActivity.java
@@ -1,30 +1,0 @@
-package com.dreamso.radio;
-
-import android.content.Intent;
-import android.os.Handler;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.view.Window;
-
-public class SplashActivity extends AppCompatActivity {
-    Handler handler;
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        getSupportActionBar().hide(); // hide the title bar
-        setContentView(R.layout.activity_splash);
-
-
-
-        handler=new Handler();
-        handler.postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                Intent intent=new Intent(SplashActivity.this,MainActivity.class);
-                startActivity(intent);
-                finish();
-            }
-        },3000);
-
-    }
-}
