--- app/src/main/java/com/dreamso/radio/player/PlaybackStatus.java
@@ -1,17 +1,0 @@
-package com.dreamso.radio.player;
-
-public class PlaybackStatus {
-
-    public static final String IDLE = "PlaybackStatus_IDLE";
-
-    public static final String LOADING = "PlaybackStatus_LOADING";
-
-    public static final String PLAYING = "PlaybackStatus_PLAYING";
-
-    public static final String PAUSED = "PlaybackStatus_PAUSED";
-
-    public static final String STOPPED = "PlaybackStatus_STOPPED";
-
-    public static final String ERROR = "PlaybackStatus_ERROR";
-
-}
