--- app/src/main/java/com/dreamso/radio/player/VisualizerView.java
@@ -1,74 +1,0 @@
-package com.dreamso.radio.player;
-
-import android.content.Context;
-import android.graphics.Canvas;
-import android.graphics.CornerPathEffect;
-import android.graphics.Paint;
-import android.graphics.Rect;
-import android.support.v4.content.ContextCompat;
-import android.util.AttributeSet;
-import android.view.View;
-
-import com.dreamso.radio.R;
-
-public class VisualizerView extends View {
-
-    private byte[] mBytes;
-    private float[] mPoints;
-    private Rect mRect = new Rect();
-    private Paint mPaint = new Paint();
-
-    public VisualizerView(Context context) {
-        super(context);
-        init();
-    }
-
-    public VisualizerView(Context context, AttributeSet attrs) {
-        super(context, attrs);
-        init();
-    }
-
-    public VisualizerView(Context context, AttributeSet attrs, int defStyleAttr) {
-        super(context, attrs, defStyleAttr);
-        init();
-    }
-
-    private void init() {
-        mBytes = null;
-        mPaint.setStrokeWidth(5.0f);
-        mPaint.setAntiAlias(true);
-        mPaint.setStyle(Paint.Style.STROKE);
-        mPaint.setStrokeJoin(Paint.Join.ROUND);
-        mPaint.setStrokeCap(Paint.Cap.ROUND);
-        mPaint.setPathEffect(new CornerPathEffect(15.0f));
-        mPaint.setColor(ContextCompat.getColor(getContext(), R.color.myorange));
-    }
-
-    public void updateVisualizer(byte[] bytes) {
-        mBytes = bytes;
-        invalidate();
-    }
-
-    @Override
-    protected void onDraw(Canvas canvas) {
-        super.onDraw(canvas);
-        if (mBytes == null) {
-            return;
-        }
-        if (mPoints == null || mPoints.length < mBytes.length * 4) {
-            mPoints = new float[mBytes.length * 4];
-        }
-        mRect.set(0, 0, getWidth(), getHeight());
-        for (int i = 0; i < mBytes.length - 1; i++) {
-            mPoints[i * 4] = mRect.width() * i / (mBytes.length - 1);
-            mPoints[i * 4 + 1] = mRect.height() / 2
-                    + ((byte) (mBytes[i] + 128)) * (mRect.height() / 2) / 128;
-            mPoints[i * 4 + 2] = mRect.width() * (i + 1) / (mBytes.length - 1);
-            mPoints[i * 4 + 3] = mRect.height() / 2
-                    + ((byte) (mBytes[i + 1] + 128)) * (mRect.height() / 2)
-                    / 128;
-        }
-        canvas.drawLines(mPoints, mPaint);
-    }
-
-}
