+++ app/src/main/java/com/dreamso/radiolk/SplashActivity.java
@@ -1,0 +1,29 @@
+package com.dreamso.radiolk;
+
+import android.content.Intent;
+import android.os.Handler;
+import androidx.appcompat.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class SplashActivity extends AppCompatActivity {
+    Handler handler;
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        getSupportActionBar().hide(); // hide the title bar
+        setContentView(R.layout.activity_splash);
+
+
+
+        handler=new Handler();
+        handler.postDelayed(new Runnable() {
+            @Override
+            public void run() {
+                Intent intent=new Intent(SplashActivity.this,MainActivity.class);
+                startActivity(intent);
+                finish();
+            }
+        },3000);
+
+    }
+}
