+++ app/src/main/java/com/dreamso/radiolk/player/MediaNotificationManager.java
@@ -1,0 +1,105 @@
+package com.dreamso.radiolk.player;
+
+import android.app.Notification;
+import android.app.NotificationChannel;
+import android.app.NotificationManager;
+import android.app.PendingIntent;
+import android.content.Context;
+import android.content.Intent;
+import android.content.res.Resources;
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+import android.os.Build;
+import androidx.core.app.NotificationCompat;
+import androidx.core.app.NotificationManagerCompat;
+
+import com.dreamso.radiolk.MainActivity;
+import com.dreamso.radiolk.R;
+
+public class MediaNotificationManager {
+
+    public static final int NOTIFICATION_ID = 555;
+    private final String PRIMARY_CHANNEL = "PRIMARY_CHANNEL_ID";
+    private final String PRIMARY_CHANNEL_NAME = "PRIMARY";
+
+    private RadioService service;
+
+    private String strAppName, strLiveBroadcast;
+
+    private Resources resources;
+
+    private NotificationManagerCompat notificationManager;
+
+    public MediaNotificationManager(RadioService service) {
+
+        this.service = service;
+        this.resources = service.getResources();
+
+        strAppName = resources.getString(R.string.app_name);
+        strLiveBroadcast = resources.getString(R.string.live_broadcast);
+
+        notificationManager = NotificationManagerCompat.from(service);
+    }
+
+    public void startNotify(String playbackStatus) {
+
+        Bitmap largeIcon = BitmapFactory.decodeResource(resources, R.drawable.radiologo);
+
+        int icon = R.drawable.ic_pause_white;
+        Intent playbackAction = new Intent(service, RadioService.class);
+        playbackAction.setAction(RadioService.ACTION_PAUSE);
+        PendingIntent action = PendingIntent.getService(service, 1, playbackAction, 0);
+
+        if(playbackStatus.equals(PlaybackStatus.PAUSED)){
+
+            icon = R.drawable.ic_play_white;
+            playbackAction.setAction(RadioService.ACTION_PLAY);
+            action = PendingIntent.getService(service, 2, playbackAction, 0);
+
+        }
+
+        Intent stopIntent = new Intent(service, RadioService.class);
+        stopIntent.setAction(RadioService.ACTION_STOP);
+        PendingIntent stopAction = PendingIntent.getService(service, 3, stopIntent, 0);
+
+        Intent intent = new Intent(service, MainActivity.class);
+        intent.setAction(Intent.ACTION_MAIN);
+        intent.addCategory(Intent.CATEGORY_LAUNCHER);
+        PendingIntent pendingIntent = PendingIntent.getActivity(service, 0, intent, 0);
+
+        notificationManager.cancel(NOTIFICATION_ID);
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            NotificationManager manager = (NotificationManager) service.getSystemService(Context.NOTIFICATION_SERVICE);
+            NotificationChannel channel = new NotificationChannel(PRIMARY_CHANNEL, PRIMARY_CHANNEL_NAME, NotificationManager.IMPORTANCE_LOW);
+            channel.setLockscreenVisibility(Notification.VISIBILITY_PRIVATE);
+            manager.createNotificationChannel(channel);
+        }
+
+        NotificationCompat.Builder builder = new NotificationCompat.Builder(service, PRIMARY_CHANNEL)
+                .setAutoCancel(false)
+                .setContentTitle(strLiveBroadcast)
+                .setContentText(strAppName)
+                .setLargeIcon(largeIcon)
+                .setContentIntent(pendingIntent)
+                .setVisibility(NotificationCompat.VISIBILITY_PUBLIC)
+                .setSmallIcon(android.R.drawable.stat_sys_headset)
+                .addAction(icon, "pause", action)
+                .addAction(R.drawable.ic_stop_white, "stop", stopAction)
+                .setPriority(NotificationCompat.PRIORITY_HIGH)
+                .setWhen(System.currentTimeMillis())
+                .setStyle(new androidx.media.app.NotificationCompat.MediaStyle()
+                        .setMediaSession(service.getMediaSession().getSessionToken())
+                        .setShowActionsInCompactView(0, 1)
+                        .setShowCancelButton(true)
+                        .setCancelButtonIntent(stopAction));
+
+        service.startForeground(NOTIFICATION_ID, builder.build());
+    }
+
+    public void cancelNotify() {
+
+        service.stopForeground(true);
+    }
+
+}
