+++ app/src/main/java/com/dreamso/radiolk/player/PlaybackStatus.java
@@ -1,0 +1,17 @@
+package com.dreamso.radiolk.player;
+
+public class PlaybackStatus {
+
+    public static final String IDLE = "PlaybackStatus_IDLE";
+
+    public static final String LOADING = "PlaybackStatus_LOADING";
+
+    public static final String PLAYING = "PlaybackStatus_PLAYING";
+
+    public static final String PAUSED = "PlaybackStatus_PAUSED";
+
+    public static final String STOPPED = "PlaybackStatus_STOPPED";
+
+    public static final String ERROR = "PlaybackStatus_ERROR";
+
+}
