+++ app/src/main/res/layout/content_main.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.dreamso.radio.MainActivity">
+    tools:context="com.dreamso.radiolk.MainActivity">
 
     <ImageView
         android:id="@+id/applogo"
@@ -191,7 +191,7 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/applogo" />
 
-    <com.dreamso.radio.player.VisualizerView
+    <com.dreamso.radiolk.player.VisualizerView
         android:id="@+id/view_visualizer"
         android:layout_width="375dp"
         android:layout_height="230dp"
@@ -211,4 +211,4 @@
         tools:ignore="MissingConstraints" />
 
 
-</android.support.constraint.ConstraintLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
