+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,7 @@
-#Tue Jul 10 12:36:06 EET 2018
+#Sun Oct 18 02:57:27 IST 2020
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 zipStoreBase=GRADLE_USER_HOME
 zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-4.4-all.zip
+//distributionUrl=https\://services.gradle.org/distributions/gradle-6.5-bin.zip
+distributionUrl=https\://services.gradle.org/distributions/gradle-4.6-all.zip
